// src/schemas/categorySchema.ts
import { z } from 'zod';
import { extendZod, zId } from "@zodyac/zod-mongoose";

// Extend zod with the zId and zUUID functions (!!important)
extendZod(z); 

// Define the schema (!!important)
export const CategorySchema = z.object({
    name: z.string().min(3).max(20).default("test"),
    //default value will be generated by mongoose
    createdAt: z.date().default(new Date()), 
    updatedAt: z.date().default(new Date()),
});


// Define the schema with the ID (!!important)
export const zCategorySchemaUdate = CategorySchema.extend({
    _id: zId(), //default value will be generated by mongoose
})

export type Category = z.infer<typeof zCategorySchemaUdate>;